[LCR-220] refactor: 상품 서비스 로직 분리
