+++ popi-item-service/src/main/java/com/lgcns/client/ManagerServiceClient.java
@@ -1,10 +1,11 @@
 package com.lgcns.client;
 
 import com.lgcns.config.FeignConfig;
-import com.lgcns.dto.popup.response.PopupInfoResponse;
+import com.lgcns.dto.response.ItemInfoResponse;
 import com.lgcns.response.SliceResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @FeignClient(
@@ -13,8 +14,10 @@ import org.springframework.web.bind.annotation.RequestParam;
         configuration = FeignConfig.class)
 public interface ManagerServiceClient {
 
-    @GetMapping("/internal/popups")
-    SliceResponse<PopupInfoResponse> findAllPopups(
-            @RequestParam(name = "lastPopupId", defaultValue = "0") Long lastPopupId,
+    @GetMapping("/internal/popups/{popupId}/items")
+    SliceResponse<ItemInfoResponse> findItemsByName(
+            @PathVariable(name = "popupId") Long popupId,
+            @RequestParam(name = "searchName", required = false) String searchName,
+            @RequestParam(name = "lastItemId", required = false) Long lastItemId,
             @RequestParam(name = "size", defaultValue = "8") int size);
 }
